chore: bug,  good seed in epoch api
